First pass of moving stuff out of UNetwork.
