+++ Assets/Mirror/Runtime/MessagePacker.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2db134099f0df4d96a84ae7a0cd9b4bc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
