+++ Assets/Mirror/Runtime/NetworkMessage.cs
@@ -1,0 +1,21 @@
+namespace Mirror
+{
+    public struct NetworkMessage
+    {
+        public short msgType;
+        public NetworkConnection conn;
+        public NetworkReader reader;
+
+        public TMsg ReadMessage<TMsg>() where TMsg : MessageBase, new()
+        {
+            TMsg msg = new TMsg();
+            msg.Deserialize(reader);
+            return msg;
+        }
+
+        public void ReadMessage<TMsg>(TMsg msg) where TMsg : MessageBase
+        {
+            msg.Deserialize(reader);
+        }
+    }
+}
