+++ Assets/Mirror/Runtime/NetworkMessage.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: eb04e4848a2e4452aa2dbd7adb801c51
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
